+++ libsolidity/interface/CompilerStack.h
@@ -176,6 +176,9 @@ public:
 		m_requestedContractNames = _contractNames;
 	}
 
+	/// Enable EVM Bytecode generation. This is enabled by default.
+	void enableEvmBytecodeGeneration(bool _enable = true) { m_generateEvmBytecode = _enable; }
+
 	/// Enable experimental generation of Yul IR code.
 	void enableIRGeneration(bool _enable = true) { m_generateIR = _enable; }
 
@@ -440,8 +443,9 @@ private:
 	langutil::EVMVersion m_evmVersion;
 	smtutil::SMTSolverChoice m_enabledSMTSolvers;
 	std::map<std::string, std::set<std::string>> m_requestedContractNames;
-	bool m_generateIR;
-	bool m_generateEwasm;
+	bool m_generateEvmBytecode = true;
+	bool m_generateIR = false;
+	bool m_generateEwasm = false;
 	std::map<std::string, util::h160> m_libraries;
 	/// list of path prefix remappings, e.g. mylibrary: github.com/ethereum = /usr/local/ethereum
 	/// "context:prefix=target"